requirements.txt
Django==5.2.4
django-crispy-forms==2.1
crispy-tailwind==0.5.0
Pillow==10.1.0
django-extensions==3.2.3
